Move id-card folder under planet folder
