+++ src/app/planet/[communityId]/components/CommunityDetail/CommunityDetail.tsx
