+++ src/app/planet/[communityId]/id-card/[userId]/components/IdCardDetail/IdCardDetail.tsx
