+++ src/app/planet/[communityId]/id-card/[userId]/components/IdCardDetail/index.ts
