+++ src/app/planet/[communityId]/id-card/[userId]/page.tsx
@@ -2,21 +2,21 @@ import 'server-only';
 
 import { commentQueryKey } from '~/api/domain/comment/comment.api';
 import { getCommentsServer } from '~/api/domain/comment/comment.api.server';
-import { CommentCount } from '~/app/id-card/[id]/components/CommentCount';
-import { CommentList } from '~/app/id-card/[id]/components/CommentList';
-import { IdCardDetail } from '~/app/id-card/[id]/components/IdCardDetail/IdCardDetail';
+import { CommentCount } from '~/app/planet/[communityId]/id-card/[userId]/components/CommentCount';
+import { CommentList } from '~/app/planet/[communityId]/id-card/[userId]/components/CommentList';
+import { IdCardDetail } from '~/app/planet/[communityId]/id-card/[userId]/components/IdCardDetail/IdCardDetail';
 import { Divider } from '~/components/Divider';
 import { HydrationProvider } from '~/components/HydrationProvider';
 import { CommentInput } from '~/modules/CommentInput';
 
 type IdCardDetailPageProps = {
   params: {
-    id: string;
+    userId: string;
   };
 };
 
-const IdCardDetailPage = async ({ params: { id } }: IdCardDetailPageProps) => {
-  const idCardId = Number(id);
+const IdCardDetailPage = async ({ params: { userId } }: IdCardDetailPageProps) => {
+  const idCardId = Number(userId);
   const pageParam = 1;
 
   const getCommentsQuery = async () => {
