+++ src/app/planet/[communityId]/page.tsx
@@ -2,19 +2,19 @@ import 'server-only';
 
 import { communityQueryKey } from '~/api/domain/community.api';
 import { getCommunityIdCardsServer } from '~/api/domain/community.api.server';
-import { CommunityDetail } from '~/app/planet/[id]/components/CommunityDetail';
-import { CommunityIdCards } from '~/app/planet/[id]/components/CommunityIdCards';
-import { IdCardCreatorButton } from '~/app/planet/[id]/components/IdCardCreatorButton';
+import { CommunityDetail } from '~/app/planet/[communityId]/components/CommunityDetail';
+import { CommunityIdCards } from '~/app/planet/[communityId]/components/CommunityIdCards';
+import { IdCardCreatorButton } from '~/app/planet/[communityId]/components/IdCardCreatorButton';
 import { HydrationProvider } from '~/components/HydrationProvider';
 
 type PlanetPageProps = {
   params: {
-    id: string;
+    communityId: string;
   };
 };
 
-const PlanetPage = async ({ params: { id } }: PlanetPageProps) => {
-  const communityId = Number(id);
+const PlanetPage = async ({ params: { communityId: communityIdParam } }: PlanetPageProps) => {
+  const communityId = Number(communityIdParam);
 
   const getCommunityIdCardsQuery = async () => {
     const data = await getCommunityIdCardsServer({ communityId });
